requirements.txt
aiosqlite==0.17.0
annotated-types==0.7.0
asyncpg==0.29.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Flask==3.0.3
iso8601==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pypika-tortoise==0.1.6
pytz==2024.1
tortoise-orm==0.21.6
typing_extensions==4.12.2
Werkzeug==3.0.4
